+++ libsolidity/codegen/YulUtilFunctions.h
@@ -24,6 +24,7 @@
 #include <liblangutil/EVMVersion.h>
 
 #include <libsolidity/ast/Types.h>
+#include <libsolidity/ast/AST.h>
 #include <libsolidity/codegen/MultiUseYulFunctionCollector.h>
 
 #include <libsolidity/interface/DebugSettings.h>
@@ -349,8 +350,17 @@ public:
 	/// @returns a function that reads a type from storage.
 	/// @param _splitFunctionTypes if false, returns the address and function signature in a
 	/// single variable.
-	std::string readFromStorage(Type const& _type, size_t _offset, bool _splitFunctionTypes);
-	std::string readFromStorageDynamic(Type const& _type, bool _splitFunctionTypes);
+	std::string readFromStorage(
+		Type const& _type,
+		size_t _offset,
+		bool _splitFunctionTypes,
+		VariableDeclaration::Location _location
+	);
+	std::string readFromStorageDynamic(
+		Type const& _type,
+		bool _splitFunctionTypes,
+		VariableDeclaration::Location _location
+	);
 
 	/// @returns a function that reads a value type from memory. Performs cleanup.
 	/// signature: (addr) -> value
@@ -376,6 +386,7 @@ public:
 	std::string updateStorageValueFunction(
 		Type const& _fromType,
 		Type const& _toType,
+		VariableDeclaration::Location _location,
 		std::optional<unsigned> const& _offset = std::optional<unsigned>()
 	);
 
@@ -496,7 +507,7 @@ public:
 	/// @returns the name of a function that will set the given storage item to
 	/// zero
 	/// signature: (slot, offset) ->
-	std::string storageSetToZeroFunction(Type const& _type);
+	std::string storageSetToZeroFunction(Type const& _type, VariableDeclaration::Location _location);
 
 	/// If revertStrings is debug, @returns the name of a function that
 	/// stores @param _message in memory position 0 and reverts.
@@ -573,8 +584,13 @@ private:
 	/// @param _splitFunctionTypes if false, returns the address and function signature in a
 	/// single variable.
 	/// @param _offset if provided, read from static offset, otherwise offset is a parameter of the Yul function.
-	std::string readFromStorageValueType(Type const& _type, std::optional<size_t> _offset, bool _splitFunctionTypes);
-
+	/// @param _location if provided, indicates whether we're reading from storage our transient storage.
+	std::string readFromStorageValueType(
+		Type const& _type,
+		std::optional<size_t> _offset,
+		bool _splitFunctionTypes,
+		VariableDeclaration::Location _location
+	);
 	/// @returns a function that reads a reference type from storage to memory (performing a deep copy).
 	std::string readFromStorageReferenceType(Type const& _type);