+++ test/cmdlineTests/ir_optimized_transient_storage_value_type/input.sol
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: GPL-3.0
+pragma solidity >=0.0.0;
+
+contract C {
+    uint transient x;
+    function f() public {
+        x = x + 1;
+    }
+}