+++ test/libsolidity/semanticTests/getters/transient_value_types.sol
@@ -8,7 +8,6 @@ contract C {
 }
 // ====
 // EVMVersion: >=cancun
-// compileViaYul: false
 // ----
 // x() -> 0
 // f() -> -1