+++ test/libsolidity/semanticTests/operators/compound_assign_transient_storage.sol
@@ -11,7 +11,6 @@ contract test {
 }
 // ====
 // EVMVersion: >=cancun
-// compileViaYul: false
 // ----
 // f(uint256,uint256): 0, 6 -> 7
 // f(uint256,uint256): 1, 3 -> 11