+++ test/libsolidity/semanticTests/operators/transient_storage_variable_increment_decrement.sol
@@ -11,6 +11,5 @@ contract C {
 }
 // ====
 // EVMVersion: >=cancun
-// compileViaYul: false
 // ----
 // f() -> 1