+++ test/libsolidity/semanticTests/variables/delete_transient_state_variable.sol
@@ -8,6 +8,5 @@ contract C {
 }
 // ====
 // EVMVersion: >=cancun
-// compileViaYul: false
 // ----
 // f() -> 0