+++ test/libsolidity/semanticTests/variables/transient_function_type_state_variable.sol
@@ -12,6 +12,5 @@ contract C {
 }
 // ====
 // EVMVersion: >=cancun
-// compileViaYul: false
 // ----
 // test() -> true