+++ test/libsolidity/semanticTests/variables/transient_state_address_variable_members.sol
@@ -10,7 +10,6 @@ contract C {
 }
 // ====
 // EVMVersion: >=cancun
-// compileViaYul: false
 // ----
 // constructor() ->
 // gas legacy: 59027