+++ test/libsolidity/semanticTests/variables/transient_state_variable_cleanup_tstore.sol
@@ -10,6 +10,5 @@ contract C {
 }
 // ====
 // EVMVersion: >=cancun
-// compileViaYul: false
 // ----
 // f() -> 0xff