+++ test/libsolidity/semanticTests/variables/transient_state_variable_tuple_assignment.sol
@@ -16,6 +16,5 @@ contract C {
 }
 // ====
 // EVMVersion: >=cancun
-// compileViaYul: false
 // ----
 // f() -> 2, 3, 4